+++ src/dom/index.ts
@@ -1,3 +1,4 @@
 export * from './download'
 export * from './copyToClipboard'
 export * from './position'
+export * from './frameLoop'